requirements.txt
llvmlite==0.43.0
numba==0.60.0
numpy==2.0.2
pygame-ce==2.5.2
scipy==1.15.2
